move-only: Move functions to blockstorage
